+++ src/parser/command_table.h
@@ -8,5 +8,6 @@ t_bool			is_between_quotes(const char *input, int reserved_char_index);
 int				get_arg_len(const char *start);
 int				get_cmd_len(const char *input);
 t_bool			is_system_command(const char *input, t_command *command);
+char			*get_redirection_position(const char *set, char *str_to_check);
 
 #endif
